+++ backend/src/auth/auth.guard.ts
@@ -2,7 +2,7 @@ import { ExecutionContext, Injectable } from '@nestjs/common';
 import { AuthGuard } from "@nestjs/passport";
 
 @Injectable()
-export class Guard42 extends AuthGuard('test') {
+export class Guard42 extends AuthGuard('OAuth2') {
 
 	async canActivate(context: ExecutionContext): Promise<any> {
 		const activate = (await super.canActivate(context)) as boolean;
