+++ backend/src/users/users.controller.ts
@@ -16,6 +16,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { tap, map } from 'rxjs/operators'
 import { join } from 'path';
 import { FriendStatus } from 'src/entities/friend-request-interface';
+import { FriendsService } from 'src/friends/friends.service';
 
 export const storage = {
   storage: diskStorage({
@@ -43,44 +44,14 @@ export class UserController {
     getUserProfile(@Req() req) {
         return req.user;
     }
-    @Get('/me/friends')
-    @ApiOperation({summary: 'Return accepted Friend Requests'})
-    async getAllFriends(@Req() req) {
-      return await this.userService.getFriends(req.user)
-    }
-
-    @Get('/me/friends-requests')
-    @ApiOperation({summary: 'Return Pending Friend Requests'})
-    async getFriendRequest(@Req() req){
-      return await this.userService.getPendingFriends(req.user)
-    }
-
-    @Get('/me/friends/add/:id')
-    @ApiOperation({summary: 'Add (id) as a friend'})
-    async addNewFriend(@Req() req, @Param('id', new ParseIntPipe()) id: number) {
-      const user = await this.userService.findById(id)
-      if (!user)
-        throw new NotFoundException('User not found')
-      return await this.userService.sendFriendRequest(req.user, id)
-    }
-
-    @Get('/me/friends/update/:id')
-    async answerFriendRequest(@Req() req, @Param('id', new ParseIntPipe()) id: number) {
-      return await this.userService.respondToRequest(req.user, id, FriendStatus.STATUS_ACCEPTED)
-    }
-
-    @Get('/me/friends/remove/:id')
-    async removeFriend(@Req() req, @Param('id', new ParseIntPipe()) id: number) {
-      return await this.userService.removeFriend(req.user, id)
-    }
   
-    @Get(':id/friends')
-    async getFriendsId(@Param('id', new ParseIntPipe()) id: number){
-      const user = await this.userService.findById(id)
-      if (!user)
-        throw new NotFoundException('User not found')
-      return await this.userService.getFriends(user)
-    }
+    // @Get(':id/friends')
+    // async getFriendsId(@Param('id', new ParseIntPipe()) id: number){
+    //   const user = await this.userService.findById(id)
+    //   if (!user)
+    //     throw new NotFoundException('User not found')
+    //   return await this.friendService.getFriends(user)
+    // }
 
     @Get(':id')
     async getUserProfileId(@Param('id', new ParseIntPipe()) id: number) {
